server/pyproject.toml
[project]
name = "tile-game-server"
version = "0.1.0"
description = "Back-end server for the daily tile game"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "pydantic-settings>=2.10.1",
    "pyyaml>=6.0.2",
    "sqlmodel>=0.0.24",
    "typer>=0.16.0",
    "uvicorn[standard]>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.1",
    "ruff>=0.12.7",
    "fastapi-cli>=0.0.8",
    "ipython>=9.4.0",
]

[tool.ruff]
line-length = 100
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "native"
